fix set Image
