+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
 
+    <string name="have_not_picked_img">Not Picked</string>
 </resources>
